+++ packages/components/psammead-media-player/src/index.stories.jsx
@@ -25,6 +25,24 @@ storiesOf('Components|Media Player', module)
     { notes, knobs: { escapeHTML: false } },
   )
   .add(
+    'Guidance',
+    () => (
+      <CanonicalMediaPlayer
+        src="https://www.test.bbc.co.uk/ws/av-embeds/articles/c3wmq4d1y3wo/p01k6msp/en"
+        placeholderSrc="https://ichef.bbci.co.uk/news/640/cpsdevpb/4eb7/test/ba7482d0-cca8-11e8-b0bf-f33155223fc4.jpg"
+        service="news"
+        title="Video player"
+        mediaInfo={{
+          title: 'Dog chases cat.',
+          guidanceMessage:
+            'Guidance: May contain strong language that may offend',
+          ...withDuration,
+        }}
+      />
+    ),
+    { notes, knobs: { escapeHTML: false } },
+  )
+  .add(
     'Audio',
     () => (
       <CanonicalMediaPlayer
@@ -45,8 +63,9 @@ storiesOf('Components|Media Player', module)
     'Audio Skin',
     () => (
       <CanonicalMediaPlayer
-        src="https://www.bbc.com/ws/av-embeds/media/bbc_korean_radio/liveradio/ko"
+        src="https://www.test.bbc.com/ws/av-embeds/media/bbc_korean_radio/liveradio/ko"
         showPlaceholder={false}
+        placeholderSrc="https://news.files.bbci.co.uk/include/articles/public/images/amp_audio_placeholder.png"
         skin="audio"
         service="news"
         mediaInfo={{ type: 'audio', title: 'Live show intro.' }}